Project/backend/requirements.txt
blinker==1.4
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Pillow==8.2.0
pycparser==2.20
PyJWT==2.0.1
six==1.15.0
Werkzeug==1.0.1
